+++ src/Composed.Commands/Compose.cs
@@ -3,6 +3,8 @@ namespace Composed.Commands
     using System;
     using System.Reactive;
     using System.Reactive.Concurrency;
+    using System.Threading;
+    using System.Threading.Tasks;
 
     /// <summary>
     ///     <para>
@@ -15,7 +17,7 @@ namespace Composed.Commands
     /// </summary>
     public static class Compose
     {
-        private static readonly CanExecuteFunc CanAlwaysExecute = () => true;
+        private static readonly Func<bool> CanAlwaysExecute = () => true;
 
         /// <summary>
         ///     Creates and returns a new <see cref="ComposedCommand"/> instance which can always
@@ -31,7 +33,7 @@ namespace Composed.Commands
         /// <exception cref="ArgumentNullException">
         ///     <paramref name="execute"/> is <see langword="null"/>.
         /// </exception>
-        public static ComposedCommand UseCommand(ExecuteAction execute)
+        public static ComposedCommand UseCommand(Action execute)
         {
             return UseCommand(execute, CanAlwaysExecute, scheduler: null, Array.Empty<IObservable<Unit>>());
         }
@@ -75,8 +77,8 @@ namespace Composed.Commands
         ///     or one of the dependencies in the <paramref name="dependencies"/> array is <see langword="null"/>.
         /// </exception>
         public static ComposedCommand UseCommand(
-            ExecuteAction execute,
-            CanExecuteFunc canExecute,
+            Action execute,
+            Func<bool> canExecute,
             params IObservable<Unit>[] dependencies
         )
         {
@@ -135,8 +137,8 @@ namespace Composed.Commands
         ///     or one of the dependencies in the <paramref name="dependencies"/> array is <see langword="null"/>.
         /// </exception>
         public static ComposedCommand UseCommand(
-            ExecuteAction execute,
-            CanExecuteFunc canExecute,
+            Action execute,
+            Func<bool> canExecute,
             IScheduler? scheduler,
             params IObservable<Unit>[] dependencies
         )
@@ -145,6 +147,166 @@ namespace Composed.Commands
             _ = canExecute ?? throw new ArgumentNullException(nameof(canExecute));
             // dependencies is ANE validated in ComposedCommand's constructor via the Watch call.
             return new ComposedCommand(execute, canExecute, scheduler, dependencies);
+        }
+
+        /// <inheritdoc cref="UseCommand(Func{CancellationToken, Task})"/>
+        public static AsyncComposedCommand UseCommand(Func<Task> executeAsync)
+        {
+            _ = executeAsync ?? throw new ArgumentNullException(nameof(executeAsync));
+            return UseCommand(_ => executeAsync());
+        }
+
+        /// <summary>
+        ///     Creates and returns a new <see cref="AsyncComposedCommand"/> instance which can always
+        ///     be executed and, when executed, invokes the specified <paramref name="executeAsync"/> function.
+        /// </summary>
+        /// <param name="executeAsync">
+        ///     The function to be invoked when the command is executed.
+        /// </param>
+        /// <returns>
+        ///     A new <see cref="AsyncComposedCommand"/> instance which can always
+        ///     be executed and, when executed, invokes the specified <paramref name="executeAsync"/> function.
+        /// </returns>
+        /// <exception cref="ArgumentNullException">
+        ///     <paramref name="executeAsync"/> is <see langword="null"/>.
+        /// </exception>
+        public static AsyncComposedCommand UseCommand(Func<CancellationToken, Task> executeAsync)
+        {
+            return UseCommand(executeAsync, CanAlwaysExecute, scheduler: null, Array.Empty<IObservable<Unit>>());
+        }
+
+        /// <inheritdoc cref="UseCommand(Func{CancellationToken, Task}, Func{bool}, IObservable{Unit}[])"/>
+        public static AsyncComposedCommand UseCommand(
+           Func<Task> executeAsync,
+           Func<bool> canExecute,
+           params IObservable<Unit>[] dependencies
+        )
+        {
+            _ = executeAsync ?? throw new ArgumentNullException(nameof(executeAsync));
+            return UseCommand(_ => executeAsync(), canExecute, dependencies);
+        }
+
+        /// <summary>
+        ///     <para>
+        ///         Creates and returns a new <see cref="AsyncComposedCommand"/> instance which invokes the
+        ///         specified <paramref name="executeAsync"/> function when it's executed and which
+        ///         uses the specified <paramref name="canExecute"/> function and <paramref name="dependencies"/>
+        ///         for determining when it can be executed.
+        ///     </para>
+        ///     <para>
+        ///         Scheduling of <paramref name="canExecute"/> invocations and any subsequent notifications
+        ///         depends on the observables passed as <paramref name="dependencies"/>.
+        ///     </para>
+        /// </summary>
+        /// <param name="executeAsync">
+        ///     The function to be invoked when the command is executed.
+        /// </param>
+        /// <param name="canExecute">
+        ///     A function which returns a boolean value indicating whether a command can be
+        ///     executed in its current state.
+        /// </param>
+        /// <param name="dependencies">
+        ///     <para>
+        ///         A set of <paramref name="canExecute"/> dependencies which will be watched for changes.
+        ///         This can be any kind of observable. <see cref="IRef{T}"/> and <see cref="IReadOnlyRef{T}"/> instances
+        ///         implement <see cref="IObservable{T}"/> and can directly be passed as dependencies.
+        ///     </para>
+        ///     <para>
+        ///         If this is empty, <paramref name="canExecute"/> will never change.
+        ///     </para>
+        /// </param>
+        /// <returns>
+        ///     A new <see cref="AsyncComposedCommand"/> instance which invokes the specified <paramref name="executeAsync"/>
+        ///     function when it's executed and which uses the specified <paramref name="canExecute"/> function and
+        ///     <paramref name="dependencies"/> for determining when it can be executed.
+        /// </returns>
+        /// <exception cref="ArgumentNullException">
+        ///     <paramref name="executeAsync"/>, <paramref name="canExecute"/>, <paramref name="dependencies"/>
+        ///     or one of the dependencies in the <paramref name="dependencies"/> array is <see langword="null"/>.
+        /// </exception>
+        public static AsyncComposedCommand UseCommand(
+            Func<CancellationToken, Task> executeAsync,
+            Func<bool> canExecute,
+            params IObservable<Unit>[] dependencies
+        )
+        {
+            return UseCommand(executeAsync, canExecute, scheduler: null, dependencies);
+        }
+
+        /// <inheritdoc cref="UseCommand(Func{CancellationToken, Task}, Func{bool}, IScheduler?, IObservable{Unit}[])"/>
+        public static AsyncComposedCommand UseCommand(
+            Func<Task> executeAsync,
+            Func<bool> canExecute,
+            IScheduler? scheduler,
+            params IObservable<Unit>[] dependencies
+        )
+        {
+            _ = executeAsync ?? throw new ArgumentNullException(nameof(executeAsync));
+            return UseCommand(_ => executeAsync(), canExecute, scheduler, dependencies);
+        }
+
+        /// <summary>
+        ///     <para>
+        ///         Creates and returns a new <see cref="AsyncComposedCommand"/> instance which invokes the
+        ///         specified <paramref name="executeAsync"/> function when it's executed and which
+        ///         uses the specified <paramref name="canExecute"/> function and <paramref name="dependencies"/>
+        ///         for determining when it can be executed.
+        ///     </para>
+        ///     <para>
+        ///         This overload allows you to specify an <see cref="IScheduler"/>
+        ///         which is used for scheduling <paramref name="canExecute"/> invocations and any subsequent
+        ///         notifications.
+        ///     </para>
+        /// </summary>
+        /// <param name="executeAsync">
+        ///     The function to be invoked when the command is executed.
+        /// </param>
+        /// <param name="canExecute">
+        ///     A function which returns a boolean value indicating whether a command can be
+        ///     executed in its current state.
+        /// </param>
+        /// <param name="scheduler">
+        ///     <para>
+        ///         An <see cref="IScheduler"/> on which the <paramref name="canExecute"/> invocations and
+        ///         any subsequent notifications are scheduled.<br/>
+        ///         <b>Important: </b> The initial invocation of <paramref name="canExecute"/> is always run immediately
+        ///         on the calling thread and <i>is not</i> scheduled on this scheduler.
+        ///     </para>
+        ///     <para>
+        ///         If this is <see langword="null"/>, scheduling of <paramref name="canExecute"/> invocations
+        ///         depends on the observables passed as <paramref name="dependencies"/>.
+        ///     </para>
+        /// </param>
+        /// <param name="dependencies">
+        ///     <para>
+        ///         A set of <paramref name="canExecute"/> dependencies which will be watched for changes.
+        ///         This can be any kind of observable. <see cref="IRef{T}"/> and <see cref="IReadOnlyRef{T}"/> instances
+        ///         implement <see cref="IObservable{T}"/> and can directly be passed as dependencies.
+        ///     </para>
+        ///     <para>
+        ///         If this is empty, <paramref name="canExecute"/> will never change.
+        ///     </para>
+        /// </param>
+        /// <returns>
+        ///     A new <see cref="AsyncComposedCommand"/> instance which invokes the specified <paramref name="executeAsync"/>
+        ///     function when it's executed and which uses the specified <paramref name="canExecute"/> function and
+        ///     <paramref name="dependencies"/> for determining when it can be executed.
+        /// </returns>
+        /// <exception cref="ArgumentNullException">
+        ///     <paramref name="executeAsync"/>, <paramref name="canExecute"/>, <paramref name="dependencies"/>
+        ///     or one of the dependencies in the <paramref name="dependencies"/> array is <see langword="null"/>.
+        /// </exception>
+        public static AsyncComposedCommand UseCommand(
+            Func<CancellationToken, Task> executeAsync,
+            Func<bool> canExecute,
+            IScheduler? scheduler,
+            params IObservable<Unit>[] dependencies
+        )
+        {
+            _ = executeAsync ?? throw new ArgumentNullException(nameof(executeAsync));
+            _ = canExecute ?? throw new ArgumentNullException(nameof(canExecute));
+            // dependencies is ANE validated in AsyncComposedCommand's constructor via the Watch call.
+            return new AsyncComposedCommand(executeAsync, canExecute, scheduler, dependencies);
         }
     }
 }